hotfix release pipeline (#165)
* fix clone code takes long time for the rest tidb scripts

* fix sh script dirs

* add script to release hotfix

* add macos script to release hotfix

* fix some components binary name do not match

* add force_rebuild trigger to adjust hotfix release

* merge same output path rule

* add force_rebuild

+++ jenkins/pipelines/cd/optimization-build-tidb-linux-arm.groovy
@@ -11,10 +11,40 @@
 * @TIFLASH_HASH
 * @RELEASE_TAG
 * @PRE_RELEASE
+* @FORCE_REBUILD
 */
 def slackcolor = 'good'
-def os = "linux"
-def arch = "arm64"
+os = "linux"
+arch = "arm64"
+platform = "centos7"
+
+def ifFileCacheExists(product,hash,binary) {
+    if (FORCE_REBUILD){
+        return false
+    }
+    if(!fileExists("gethash.py")){
+        sh "curl -s ${FILE_SERVER_URL}/download/builds/pingcap/ee/gethash.py > gethash.py"
+    }
+    def filepath = "builds/pingcap/${product}/optimization/${hash}/${platform}/${binary}-${os}-${arch}.tar.gz"
+    if (product == "br") {
+        filepath = "builds/pingcap/${product}/optimization/${RELEASE_TAG}/${hash}/${platform}/${binary}-${os}-${arch}.tar.gz"
+    }
+    if (product == "ticdc") {
+        filepath = "builds/pingcap/${product}/optimization/${hash}/${platform}/${product}-${os}-${arch}.tar.gz"
+    }
+    if (product == "tiflash") {
+        filepath = "builds/pingcap/${product}/optimization/${RELEASE_TAG}/${hash}/${platform}/${binary}-${os}-${arch}.tar.gz"
+    }
+
+    result = sh(script: "curl -I ${FILE_SERVER_URL}/download/${filepath} -X \"HEAD\"|grep \"200 OK\"", returnStatus: true)
+    // result equal 0 mean cache file exists
+    if (result==0) {
+        echo "file ${FILE_SERVER_URL}/download/${filepath} found in cache server,skip build again"
+        return true
+    }
+    return false
+}
+
 def TIDB_CTL_HASH = "master"
 def build_upload = { product, hash, binary ->
     stage("Build ${product}") {
@@ -131,29 +161,45 @@ try {
         builds = [:]
 
         if (BUILD_TIKV_IMPORTER == "false") {
-            builds["Build tidb-ctl"] = {
+            if (!ifFileCacheExists("tidb-ctl", TIDB_CTL_HASH, "tidb-ctl")) {
+                builds["Build tidb-ctl"] = {
                 build_upload("tidb-ctl", TIDB_CTL_HASH, "tidb-ctl")
+                } 
             }
-            builds["Build tidb"] = {
+            if (!ifFileCacheExists("tidb", TIDB_HASH, "tidb-server")) {
+                builds["Build tidb"] = {
                 build_upload("tidb", TIDB_HASH, "tidb-server")
+                }
             }
-            builds["Build tidb-binlog"] = {
+            if (!ifFileCacheExists("tidb-binlog", BINLOG_HASH, "tidb-binlog")) {
+                builds["Build tidb-binlog"] = {
                 build_upload("tidb-binlog", BINLOG_HASH, "tidb-binlog")
+                }
             }
-            builds["Build tidb-tools"] = {
+            if (!ifFileCacheExists("tidb-tools", TOOLS_HASH, "tidb-tools")) {
+                builds["Build tidb-tools"] = {
                 build_upload("tidb-tools", TOOLS_HASH, "tidb-tools")
+                }
             }
-            builds["Build pd"] = {
+            if (!ifFileCacheExists("pd", PD_HASH, "pd-server")) {
+                builds["Build pd"] = {
                 build_upload("pd", PD_HASH, "pd-server")
+                }
             }
-            builds["Build ticdc"] = {
+            if (!ifFileCacheExists("ticdc", CDC_HASH, "ticdc")) {
+                builds["Build ticdc"] = {
                 build_upload("ticdc", CDC_HASH, "ticdc")
+                }
             }
-            builds["Build br"] = {
+            if (!ifFileCacheExists("br", BR_HASH, "br")) {
+                builds["Build br"] = {
                 build_upload("br", BR_HASH, "br")
+                }
             }
-            builds["Build dumpling"] = {
+            if (!ifFileCacheExists("dumpling", DUMPLING_HASH, "dumpling")) {
+                builds["Build dumpling"] = {
                 build_upload("dumpling", DUMPLING_HASH, "dumpling")
+                }
             }
         }
         if (SKIP_TIFLASH == "false" && BUILD_TIKV_IMPORTER == "false") {
@@ -170,6 +216,9 @@ try {
                         container("tiflash-build-arm") {
                             // println "arm debug command:\nkubectl -n jenkins-ci exec -ti ${NODE_NAME} bash"
                             dir("tics") {
+                                if (!ifFileCacheExists("tiflash", TIFLASH_HASH, "tiflash")) {
+                                    return
+                                }
                                 def target = "tiflash-${RELEASE_TAG}-${os}-${arch}"
                                 def filepath = "builds/pingcap/tiflash/optimization/${RELEASE_TAG}/${TIFLASH_HASH}/centos7/tiflash-${os}-${arch}.tar.gz"
                                 try {
@@ -227,6 +276,9 @@ try {
         builds["Build tikv"] = {
             node("arm") {
                 dir("go/src/github.com/pingcap/tikv") {
+                    if (!ifFileCacheExists("tikv", TIKV_HASH, "tikv-server")) {
+                        return
+                    }
                     if (sh(returnStatus: true, script: '[ -d .git ] || git rev-parse --git-dir > /dev/null 2>&1') != 0) {
                         deleteDir()
                     }
@@ -268,6 +320,9 @@ try {
         builds["Build importer"] = {
             node("arm") {
                 dir("go/src/github.com/pingcap/importer") {
+                    if (!ifFileCacheExists("importer", IMPORTER_HASH, "importer")) {
+                        return
+                    }
                     def target = "importer-${RELEASE_TAG}-${os}-${arch}"
                     def filepath = "builds/pingcap/importer/optimization/${IMPORTER_HASH}/centos7/importer-${os}-${arch}.tar.gz"
                     try {
